+++ src/events/process_event_acknowledge_packet.py
@@ -1,4 +1,4 @@
-from src.mutations.events import create_event_acknowledge_packet_mutation
+from src.mutations.create_events import create_event_acknowledge_packet_mutation
 
 
 def process_event_acknowledge_packet_event(hash, event_type, events, height, timestamp):
