+++ src/events/process_event_add_liquidity.py
@@ -1,6 +1,6 @@
 import re
 
-from src.mutations.events import create_event_add_liquidity_mutation
+from src.mutations.create_events import *
 from src.utils.clean_parse_amount import clean_parse_amount_util
 from src.utils.clean_parse_token import clean_parse_token_util
