+++ src/events/process_event_distribution_started.py
@@ -1,4 +1,4 @@
-from src.mutations.create_event_distribution_started import create_event_distribution_started_mutation
+from src.mutations.create_events import create_event_distribution_started_mutation
 from src.utils.clean_parse_amount import clean_parse_amount_util
 from src.utils.clean_parse_token import clean_parse_token_util
