+++ src/events/process_event_edit_validator.py
@@ -1,6 +1,6 @@
 import re
 
-from src.mutations.create_event_edit_validator import create_event_edit_validator_mutation
+from src.mutations.create_events import create_event_edit_validator_mutation
 
 
 def process_event_edit_validator_event(hash, event_type, events, height, timestamp, tx):
