+++ src/events/process_event_ibc_transfer.py
@@ -1,4 +1,4 @@
-from src.mutations.create_event_denomination_trace import create_event_denomination_trace_mutation
+from src.mutations.create_events import create_event_denomination_trace_mutation
 from src.utils.clean_parse_amount import clean_parse_amount_util
 from src.utils.clean_parse_token import clean_parse_token_util
