+++ src/events/process_event_proposal_deposit.py
@@ -1,4 +1,4 @@
-from src.mutations.create_event_proposal_deposit import create_event_proposal_deposit_mutation
+from src.mutations.create_events import create_event_proposal_deposit_mutation
 from src.utils.clean_parse_amount import clean_parse_amount_util
 from src.utils.clean_parse_token import clean_parse_token_util
