+++ src/events/process_event_record_withdraw_rewards.py
@@ -1,6 +1,6 @@
 import re
 
-from src.mutations.create_event_withdraw_rewards import create_event_withdraw_rewards_mutation
+from src.mutations.create_events import create_event_withdraw_rewards_mutation
 from src.utils.clean_parse_amount import clean_parse_amount_util
