+++ src/events/process_event_remove_liquidity.py
@@ -1,7 +1,7 @@
 import logging
 import re
 
-from src.mutations.create_event_remove_liquidity import create_event_remove_liquidity_mutation
+from src.mutations.create_events import create_event_remove_liquidity_mutation
 from src.utils.clean_parse_amount import clean_parse_amount_util
 from src.utils.clean_parse_token import clean_parse_token_util
 from src.utils.setup_logger import setup_logger_util
