+++ src/events/process_event_unknown.py
@@ -1,4 +1,4 @@
-from src.mutations.create_event_unknown import create_event_unknown_mutation
+from src.mutations.create_events import create_event_unknown_mutation
 
 
 def process_event_unknown_event(hash, event_type, events, height, timestamp):
