--- src/mutations/create_event_proposal_vote.py
@@ -1,20 +1,0 @@
-import json
-
-from src.services.database import database_service
-
-
-def create_event_proposal_vote_mutation(hash, event_type,
-                                        events_arr, height, timestamp,
-                                        sender, vote, gasWanted, gasUsed):
-
-    sql_str = '''
-    INSERT INTO events_audit
-    (hash, type, log, height, time,
-    pv_sender, pv_vote, pv_gasWanted, pv_gasUsed)
-    VALUES ('{0}', '{1}', '{2}', '{3}', '{4}', 
-    '{5}', '{6}', '{7}', '{8}' 
-    )
-    '''.format(hash, event_type, json.dumps(events_arr), height, timestamp,
-               sender, vote, gasWanted, gasUsed)
-
-    database_service.execute_update(sql_str)
